Fix cb::RingBuffer iterators
Currently the RingBuffer iterators can't be used in a generic
fashion across the standard library because they don't define the
required member types. This commit adds the base set of types.

Also fixes a build warning in the RingBuffer test.

Reviewed-on: http://review.couchbase.org/70959
Tested-by: buildbot <build@couchbase.com>

+++ tests/ring_buffer/ring_buffer_test.cc
@@ -89,7 +89,7 @@ TYPED_TEST(RingBufferTest, testRingBuffer) {
     }
 
     ASSERT_EQ(10, rb.size());
-    for (int i = 0; i < rb.size(); ++i) {
+    for (size_t i = 0; i < rb.size(); ++i) {
         ASSERT_EQ(i + 1, rb[i]);
     }
 }
